style: refactor coe
